Moved styles from Dashboard to Styles folder
