+++ client/src/components/Styles/Dashboard/styles.js
@@ -1,0 +1,26 @@
+const styles = theme => ({
+  root: {
+    flexGrow: 1,
+    backgroundColor: theme.palette.grey["100"],
+    overflow: "hidden"
+    //height: "100vh"
+  },
+  grid: {
+    width: 1200,
+    margin: `0 ${theme.spacing.unit * 2}px`,
+    [theme.breakpoints.down("sm")]: {
+      width: "calc(100% - 20px)"
+    }
+  },
+  paper: {
+    padding: theme.spacing.unit * 3,
+    textAlign: "left",
+    color: theme.palette.text.secondary
+    // height: '40vh',
+  },
+  block: {
+    padding: theme.spacing.unit * 2
+  }
+});
+
+export default styles;
