+++ client/src/components/Styles/Reports/styles.js
@@ -1,5 +1,3 @@
-import { fade } from "@material-ui/core/styles/colorManipulator";
-
 const styles = theme => ({
     root: {
       flexGrow: 1,
